requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
httpx==0.27.0
openpyxl==3.1.5
xlrd==2.0.1
beautifulsoup4==4.12.3
pydantic==2.8.0
python-dotenv==1.0.1
